be/requirements.txt
Django==4.1
djangorestframework==3.13.1
python-dotenv==0.20.0
django-cors-headers==3.13.0
django-base64field==1.0
django-filter==21.1
numpy==1.23.5